+++ src/pages/kitchensink/Kitchensink.js
@@ -5,7 +5,7 @@ import ButtonExample from "./ButtonExample";
 import ExpandableSectionExample from "./ExpandableSectionExample";
 import TwoWayDatabingExample from "./TwoWayDatabingExample";
 import ListExample from "./ListExample";
-import { BackendApi } from "./BackendApi";
+import BackendApi from "./BackendApi";
 
 class Kitchensink extends React.Component {